Added code for the segmented control and button title in the Sign Up view.
